abstract out the mem_writer, we are getting close to get rid of salls's code, amazing!
